Fix lint/fix scripts for backend
The .prettierignore was being picked up by eslint, causing eslint
formatting of ECMAScript files to break. Renaming the file to
`.other.prettierignore` prevents this.

Use the newly named ignore file only for the prettier command in the
scripts (`lint.other` and `fix.other`).

To improve performance, install and use `npm-run-all` to parallelize
the lint scripts.

Finally, add the missing prettier-plugin-solidity dep to backend/. It
was available in the parent project's package.json, but that's not
correct.